+++ src/test/java/com/iexec/sms/tee/config/TeeWorkerPipelineConfigurationTests.java
@@ -41,40 +41,54 @@ class TeeWorkerPipelineConfigurationTests {
     private static final String POST_FINGERPRINT = "postFingerprint";
     private static final DataSize POST_HEAP_SIZE = DataSize.parse("3GB");
     private static final String POST_ENTRYPOINT = "postEntrypoint";
+    private final TeeWorkerPipelineConfiguration.StageConfig preCompute = new TeeWorkerPipelineConfiguration.StageConfig(
+            PRE_IMAGE,
+            PRE_FINGERPRINT,
+            PRE_HEAP_SIZE,
+            PRE_ENTRYPOINT);
+    private final TeeWorkerPipelineConfiguration.StageConfig postCompute = new TeeWorkerPipelineConfiguration.StageConfig(
+            POST_IMAGE,
+            POST_FINGERPRINT,
+            POST_HEAP_SIZE,
+            POST_ENTRYPOINT);
 
     private Validator validator;
 
     @BeforeEach
     void setUp() {
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        final ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
         validator = factory.getValidator();
     }
 
     @Test
     void shouldGetPipelineConfiguration() {
-        final TeeWorkerPipelineConfiguration config = createValidConfiguration();
+        final TeeWorkerPipelineConfiguration.Pipeline pipeline = new TeeWorkerPipelineConfiguration.Pipeline(
+                VERSION,
+                preCompute,
+                postCompute
+        );
+        final TeeWorkerPipelineConfiguration config = new TeeWorkerPipelineConfiguration(Collections.singletonList(pipeline));
 
         assertEquals(1, config.getPipelines().size());
         final TeeWorkerPipelineConfiguration.Pipeline firstPipeline = config.getPipelines().get(0);
         assertEquals(VERSION, firstPipeline.getVersion());
 
         final TeeWorkerPipelineConfiguration.StageConfig preComputeConfig = firstPipeline.getPreCompute();
-        assertEquals(PRE_IMAGE, preComputeConfig.getImage());
-        assertEquals(PRE_FINGERPRINT, preComputeConfig.getFingerprint());
-        assertEquals(PRE_HEAP_SIZE, preComputeConfig.getHeapSize());
-        assertEquals(PRE_ENTRYPOINT, preComputeConfig.getEntrypoint());
+        assertEquals(PRE_IMAGE, preComputeConfig.image());
+        assertEquals(PRE_FINGERPRINT, preComputeConfig.fingerprint());
+        assertEquals(PRE_HEAP_SIZE, preComputeConfig.heapSize());
+        assertEquals(PRE_ENTRYPOINT, preComputeConfig.entrypoint());
 
         final TeeWorkerPipelineConfiguration.StageConfig postComputeConfig = firstPipeline.getPostCompute();
-        assertEquals(POST_IMAGE, postComputeConfig.getImage());
-        assertEquals(POST_FINGERPRINT, postComputeConfig.getFingerprint());
-        assertEquals(POST_HEAP_SIZE, postComputeConfig.getHeapSize());
-        assertEquals(POST_ENTRYPOINT, postComputeConfig.getEntrypoint());
+        assertEquals(POST_IMAGE, postComputeConfig.image());
+        assertEquals(POST_FINGERPRINT, postComputeConfig.fingerprint());
+        assertEquals(POST_HEAP_SIZE, postComputeConfig.heapSize());
+        assertEquals(POST_ENTRYPOINT, postComputeConfig.entrypoint());
     }
 
     @Test
-    void shouldValidateEmptyPipelineList() {
-        final TeeWorkerPipelineConfiguration config = new TeeWorkerPipelineConfiguration(Collections.emptyList());
-        final Set<ConstraintViolation<TeeWorkerPipelineConfiguration>> violations = validator.validate(config);
+    void shouldFailWhenPipelineConfigIsNull() {
+        final Set<ConstraintViolation<TeeWorkerPipelineConfiguration>> violations = validator.validate(new TeeWorkerPipelineConfiguration(null));
         assertFalse(violations.isEmpty());
         assertThat(violations)
                 .extracting(ConstraintViolation::getMessage)
@@ -82,9 +96,18 @@ class TeeWorkerPipelineConfigurationTests {
     }
 
     @Test
-    void shouldValidateNullPipeline() {
-        final TeeWorkerPipelineConfiguration config = new TeeWorkerPipelineConfiguration(Collections.singletonList(null));
-        final Set<ConstraintViolation<TeeWorkerPipelineConfiguration>> violations = validator.validate(config);
+    void shouldFailIfEmptyPipelineList() {
+        final Set<ConstraintViolation<TeeWorkerPipelineConfiguration>> violations = validator.validate(new TeeWorkerPipelineConfiguration(Collections.emptyList()));
+        assertFalse(violations.isEmpty());
+        assertThat(violations)
+                .extracting(ConstraintViolation::getMessage)
+                .containsExactly("Pipeline list must not be empty");
+    }
+
+    @Test
+    void shouldFailWhenListElementsAreNull() {
+        final Set<ConstraintViolation<TeeWorkerPipelineConfiguration>> violations = validator.validate(
+                new TeeWorkerPipelineConfiguration(Collections.singletonList(null)));
         assertFalse(violations.isEmpty());
         assertThat(violations)
                 .extracting(ConstraintViolation::getMessage)
@@ -92,21 +115,21 @@ class TeeWorkerPipelineConfigurationTests {
     }
 
     @Test
-    void shouldValidateBlankVersion() {
+    void shouldFailIfBlankVersion() {
         final TeeWorkerPipelineConfiguration.Pipeline pipeline = new TeeWorkerPipelineConfiguration.Pipeline(
                 "",
-                createValidPreComputeConfig(),
-                createValidPostComputeConfig()
+                preCompute,
+                postCompute
         );
         final Set<ConstraintViolation<TeeWorkerPipelineConfiguration.Pipeline>> violations = validator.validate(pipeline);
         assertFalse(violations.isEmpty());
         assertThat(violations)
                 .extracting(ConstraintViolation::getMessage)
-                .containsExactly("Pipeline version must be provided");
+                .containsExactly("Pipeline version must not be blank");
     }
 
     @Test
-    void shouldValidateNullStageConfig() {
+    void shouldFailIfNullStageConfig() {
         final TeeWorkerPipelineConfiguration.Pipeline pipeline = new TeeWorkerPipelineConfiguration.Pipeline(
                 VERSION,
                 null,
@@ -122,7 +145,7 @@ class TeeWorkerPipelineConfigurationTests {
     }
 
     @Test
-    void shouldValidateStageConfigFields() {
+    void shouldFailIfStageConfigFieldsAreInvalid() {
         final TeeWorkerPipelineConfiguration.StageConfig config = new TeeWorkerPipelineConfiguration.StageConfig(
                 "",
                 "",
@@ -133,36 +156,10 @@ class TeeWorkerPipelineConfigurationTests {
         assertEquals(4, violations.size());
         assertThat(violations)
                 .extracting(ConstraintViolation::getMessage)
-                .containsExactlyInAnyOrder("Fingerprint must not be blank",
+                .containsExactlyInAnyOrder(
+                        "Fingerprint must not be blank",
                         "Image must not be blank",
                         "Heap size must not be null",
                         "Entrypoint must not be blank");
-    }
-
-    private TeeWorkerPipelineConfiguration createValidConfiguration() {
-        final TeeWorkerPipelineConfiguration.Pipeline pipeline = new TeeWorkerPipelineConfiguration.Pipeline(
-                VERSION,
-                createValidPreComputeConfig(),
-                createValidPostComputeConfig()
-        );
-        return new TeeWorkerPipelineConfiguration(Collections.singletonList(pipeline));
-    }
-
-    private TeeWorkerPipelineConfiguration.StageConfig createValidPreComputeConfig() {
-        return new TeeWorkerPipelineConfiguration.StageConfig(
-                PRE_IMAGE,
-                PRE_FINGERPRINT,
-                PRE_HEAP_SIZE,
-                PRE_ENTRYPOINT
-        );
-    }
-
-    private TeeWorkerPipelineConfiguration.StageConfig createValidPostComputeConfig() {
-        return new TeeWorkerPipelineConfiguration.StageConfig(
-                POST_IMAGE,
-                POST_FINGERPRINT,
-                POST_HEAP_SIZE,
-                POST_ENTRYPOINT
-        );
     }
 }